Small tweaks in TestGLFW_TextureArrayAnimation. GetFrameTime() is expressed in seconds

+++ libraries/chaos/src/BitmapAtlas.cpp
@@ -295,7 +295,7 @@ namespace chaos
 			return (size_t)MathTools::Maximum((animation_info->grid_data.grid_size.x * animation_info->grid_data.grid_size.y) - animation_info->grid_data.skip_lasts, 0);
 		}
 
-		int BitmapInfo::GetFrameTime() const
+		double BitmapInfo::GetFrameTime() const
 		{
 			if (animation_info == nullptr)
 				return 0;
